+++ src/blocks/tipset.rs
@@ -3,6 +3,7 @@
 
 use std::{fmt, sync::OnceLock};
 
+use crate::db::{SettingsStore, SettingsStoreExt};
 use crate::ipld::FrozenCids;
 use crate::networks::{calibnet, mainnet};
 use crate::shim::{address::Address, clock::ChainEpoch};
@@ -149,6 +150,25 @@ impl Tipset {
             .collect::<anyhow::Result<Option<_>>>()?
             .map(Tipset::new)
             .transpose()?)
+    }
+
+    /// Load the heaviest tipset from the blockstore
+    pub fn load_heaviest(
+        store: &impl Blockstore,
+        settings: &impl SettingsStore,
+    ) -> anyhow::Result<Option<Tipset>> {
+        Ok(
+            match settings.read_obj::<TipsetKeys>(crate::db::setting_keys::HEAD_KEY)? {
+                Some(tsk) => tsk
+                    .cids
+                    .into_iter()
+                    .map(|key| BlockHeader::load(store, key))
+                    .collect::<anyhow::Result<Option<_>>>()?
+                    .map(Tipset::new)
+                    .transpose()?,
+                None => None,
+            },
+        )
     }
 
     /// Fetch a tipset from the blockstore. This calls fails if the tipset is