+++ src/db/car/many.rs
@@ -9,19 +9,19 @@
 //! A single z-frame cache is shared between all read-only stores.
 
 use super::{AnyCar, ZstdFrameCache};
-use crate::db::MemoryDB;
+use crate::db::{MemoryDB, SettingsStore};
 use crate::libp2p_bitswap::BitswapStoreReadWrite;
+use crate::utils::io::EitherMmapOrRandomAccessFile;
 use crate::{blocks::Tipset, libp2p_bitswap::BitswapStoreRead};
 use anyhow::Context;
 use cid::Cid;
 use fvm_ipld_blockstore::Blockstore;
-use parking_lot::Mutex;
-use positioned_io::RandomAccessFile;
+use parking_lot::{Mutex, RwLock};
 use std::{io, path::PathBuf, sync::Arc};
 
 pub struct ManyCar<WriterT = MemoryDB> {
     shared_cache: Arc<Mutex<ZstdFrameCache>>,
-    read_only: Vec<AnyCar<Box<dyn super::RandomAccessFileReader>>>,
+    read_only: RwLock<Vec<AnyCar<Box<dyn super::RandomAccessFileReader>>>>,
     writer: WriterT,
 }
 
@@ -29,7 +29,7 @@ impl<WriterT> ManyCar<WriterT> {
     pub fn new(writer: WriterT) -> Self {
         ManyCar {
             shared_cache: Arc::new(Mutex::new(ZstdFrameCache::default())),
-            read_only: Vec::new(),
+            read_only: RwLock::new(Vec::new()),
             writer,
         }
     }
@@ -46,26 +46,41 @@ impl<WriterT: Default> Default for ManyCar<WriterT> {
 }
 
 impl<WriterT> ManyCar<WriterT> {
-    pub fn read_only<ReaderT: super::RandomAccessFileReader>(&mut self, any_car: AnyCar<ReaderT>) {
-        let key = self.read_only.len() as u64;
-        self.read_only.push(
+    pub fn with_read_only<ReaderT: super::RandomAccessFileReader>(
+        self,
+        any_car: AnyCar<ReaderT>,
+    ) -> Self {
+        self.read_only(any_car);
+        self
+    }
+
+    pub fn read_only<ReaderT: super::RandomAccessFileReader>(&self, any_car: AnyCar<ReaderT>) {
+        let mut read_only = self.read_only.write();
+        let key = read_only.len() as u64;
+        read_only.push(
             any_car
                 .with_cache(self.shared_cache.clone(), key)
                 .into_dyn(),
         );
     }
 
-    pub fn read_only_files(&mut self, files: impl Iterator<Item = PathBuf>) -> io::Result<()> {
+    pub fn with_read_only_files(self, files: impl Iterator<Item = PathBuf>) -> io::Result<Self> {
+        self.read_only_files(files)?;
+        Ok(self)
+    }
+
+    pub fn read_only_files(&self, files: impl Iterator<Item = PathBuf>) -> io::Result<()> {
         for file in files {
-            let car = AnyCar::new(RandomAccessFile::open(file)?)?;
-            self.read_only(car);
+            self.read_only(AnyCar::new(EitherMmapOrRandomAccessFile::open(file)?)?);
         }
+
         Ok(())
     }
 
     pub fn heaviest_tipset(&self) -> anyhow::Result<Tipset> {
         let tipsets = self
             .read_only
+            .read()
             .iter()
             .map(AnyCar::heaviest_tipset)
             .collect::<anyhow::Result<Vec<_>>>()?;
@@ -78,18 +93,14 @@ impl<WriterT> ManyCar<WriterT> {
 
 impl<ReaderT: super::RandomAccessFileReader> From<AnyCar<ReaderT>> for ManyCar<MemoryDB> {
     fn from(any_car: AnyCar<ReaderT>) -> Self {
-        let mut many_car = ManyCar::default();
-        many_car.read_only(any_car);
-        many_car
+        ManyCar::default().with_read_only(any_car)
     }
 }
 
 impl TryFrom<Vec<PathBuf>> for ManyCar<MemoryDB> {
     type Error = io::Error;
     fn try_from(files: Vec<PathBuf>) -> io::Result<Self> {
-        let mut many_car = ManyCar::default();
-        many_car.read_only_files(files.into_iter())?;
-        Ok(many_car)
+        ManyCar::default().with_read_only_files(files.into_iter())
     }
 }
 
@@ -98,7 +109,7 @@ impl<WriterT: Blockstore> Blockstore for ManyCar<WriterT> {
         // Theoretically it should be easily parallelizable with `rayon`.
         // In practice, there is a massive performance loss when providing
         // more than a single reader.
-        for reader in self.read_only.iter() {
+        for reader in self.read_only.read().iter() {
             if let Some(val) = reader.get(k)? {
                 return Ok(Some(val));
             }
@@ -129,6 +140,24 @@ impl<WriterT: BitswapStoreReadWrite + Blockstore> BitswapStoreReadWrite for Many
     }
 }
 
+impl<WriterT: SettingsStore> SettingsStore for ManyCar<WriterT> {
+    fn read_bin(&self, key: &str) -> anyhow::Result<Option<Vec<u8>>> {
+        SettingsStore::read_bin(self.writer(), key)
+    }
+
+    fn write_bin(&self, key: &str, value: &[u8]) -> anyhow::Result<()> {
+        SettingsStore::write_bin(self.writer(), key, value)
+    }
+
+    fn exists(&self, key: &str) -> anyhow::Result<bool> {
+        SettingsStore::exists(self.writer(), key)
+    }
+
+    fn setting_keys(&self) -> anyhow::Result<Vec<String>> {
+        SettingsStore::setting_keys(self.writer())
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::super::AnyCar;
@@ -143,9 +172,9 @@ mod tests {
 
     #[test]
     fn many_car_idempotent() {
-        let mut many = ManyCar::new(MemoryDB::default());
-        many.read_only(AnyCar::try_from(mainnet::DEFAULT_GENESIS).unwrap());
-        many.read_only(AnyCar::try_from(mainnet::DEFAULT_GENESIS).unwrap());
+        let many = ManyCar::new(MemoryDB::default())
+            .with_read_only(AnyCar::try_from(mainnet::DEFAULT_GENESIS).unwrap())
+            .with_read_only(AnyCar::try_from(mainnet::DEFAULT_GENESIS).unwrap());
         assert_eq!(
             many.heaviest_tipset().unwrap(),
             AnyCar::try_from(mainnet::DEFAULT_GENESIS)