+++ src/db/migration/mod.rs
@@ -1,0 +1,8 @@
+// Copyright 2019-2023 ChainSafe Systems
+// SPDX-License-Identifier: Apache-2.0, MIT
+
+mod db_migration;
+mod migration_map;
+mod v0_12_1;
+
+pub use db_migration::DbMigration;