+++ src/tool/main.rs
@@ -22,10 +22,12 @@ where
             // Run command
             match cmd {
                 Subcommand::Benchmark(cmd) => cmd.run().await,
+                Subcommand::StateMigration(state_migration) => state_migration.run().await,
                 Subcommand::Snapshot(cmd) => cmd.run().await,
                 Subcommand::Fetch(cmd) => cmd.run().await,
                 Subcommand::Archive(cmd) => cmd.run().await,
                 Subcommand::DB(cmd) => cmd.run().await,
+                Subcommand::Car(cmd) => cmd.run().await,
             }
         })
 }